// Frontend/src/swagger/model/PostOrderDto.ts
/**
 * BikeStore
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import * as models from './models';

export interface PostOrderDto {
    customerId?: number;

    storeId?: number;

    staffId?: number;

    orderStatus?: number;

    orderDate?: string;

    requiredDate?: string;

}